+++ spec/models/membership_spec.rb
@@ -173,6 +173,24 @@ RSpec.describe Membership, type: :model do
     let(:msg) { instance_spy(ActionMailer::MessageDelivery, deliver_later: 1) }
 
     # rubocop:disable RSpec/ExampleLength
+    it 'emails leader on request creation' do
+      g = create(:open_group)
+      m = Membership.new(user: create(:student, draw: g.draw), group: g,
+                         status: 'requested')
+      allow(StudentMailer).to receive(:requested_to_join_group).and_return(msg)
+      m.save
+      expect(StudentMailer).to have_received(:requested_to_join_group)
+    end
+
+    it 'emails student on invitation creation' do
+      g = create(:open_group)
+      m = Membership.new(user: create(:student, draw: g.draw), group: g,
+                         status: 'invited')
+      allow(StudentMailer).to receive(:invited_to_join_group).and_return(msg)
+      m.save
+      expect(StudentMailer).to have_received(:invited_to_join_group)
+    end
+
     it 'emails leader on invitation acceptance' do
       g = FactoryGirl.create(:open_group)
       m = Membership.create(user: FactoryGirl.create(:student, draw: g.draw),
@@ -181,7 +199,17 @@ RSpec.describe Membership, type: :model do
       m.update(status: 'accepted')
       expect(StudentMailer).to have_received(:joined_group)
     end
+
+    it 'does not email leaders on request acceptance' do
+      g = FactoryGirl.create(:open_group)
+      m = Membership.create(user: FactoryGirl.create(:student, draw: g.draw),
+                            group: g, status: 'requested')
+      allow(StudentMailer).to receive(:joined_group).and_return(msg)
+      m.update(status: 'accepted')
+      expect(StudentMailer).not_to have_received(:joined_group)
+    end
     # rubocop:enable RSpec/ExampleLength
+
     it 'emails leader when someone leaves' do
       group = FactoryGirl.create(:full_group)
       m = group.memberships.last